refactor(runtime): switch to public API imports (14 violations fixed)
Replace 14 internal module imports with public API imports:
- config.orchestra_config.PeriodicCheckConfig → config.PeriodicCheckConfig
- models.orchestra_config.OrchestraConfig → models.OrchestraConfig
- orchestra.logging.append_orchestra_event → orchestra.append_orchestra_event
- orchestra.logging.append_orchestra_run_separator → orchestra.append_orchestra_run_separator
- services.error_tracking_service.ErrorTrackingService → services.ErrorTrackingService
- services.check_service.CheckService → services.CheckService
- clients.git_client.GitClient → clients.GitClient
- clients.github_client.GitHubClient → clients.GitHubClient
- clients.sqlite_client.SQLiteClient → clients.SQLiteClient

Remove 5 external re-exports from runtime/__init__.py:
- PeriodicCheckConfig, OrchestraConfig
- append_orchestra_event, append_orchestra_run_separator
- ErrorTrackingService

Documented exception: ExpiredResourceCleanupService (not in services public API)

Verification:
- 51 tests passed (runtime + failed_gate integration)
- mypy: no errors
- ruff: all checks passed
- Zero external consumers of removed re-exports
