+++ src/vibe3/runtime/__init__.py
@@ -2,25 +2,11 @@
 
 Public interface for runtime components including circuit breaker,
 heartbeat server, periodic check executors, and orchestra instance management.
-
-This module serves as the unified public interface for runtime components.
-Import from vibe3.runtime instead of external modules to ensure clean dependencies.
 """
 
 from typing import TYPE_CHECKING
 
-from vibe3.config.orchestra_config import PeriodicCheckConfig
-
-# Module-level re-exports from models/ and config/ (safe, no cycle risk)
-from vibe3.models.orchestra_config import OrchestraConfig
-
-# Module-level re-exports from orchestra.logging (safe, no cycle risk)
-from vibe3.orchestra.logging import (
-    append_orchestra_event,
-    append_orchestra_run_separator,
-)
-
-# Lazy imports via __getattr__ for everything else to avoid circular dependencies
+# Lazy imports via __getattr__ for everything to avoid circular dependencies
 if TYPE_CHECKING:
     from vibe3.runtime.circuit_breaker import (
         CircuitBreaker,
@@ -38,14 +24,12 @@ if TYPE_CHECKING:
     )
     from vibe3.runtime.periodic_check_executor import execute_periodic_check
     from vibe3.runtime.service_protocol import ServiceBase
-    from vibe3.services.error_tracking_service import ErrorTrackingService
 
 
 def __getattr__(name: str) -> object:
     """Lazy import for all symbols to avoid circular dependencies.
 
-    All symbols (runtime submodules, services) are imported lazily
-    to prevent circular import issues.
+    All symbols are imported lazily to prevent circular import issues.
     """
     # Orchestra instance management
     if name == "OrchestraInstanceInfo":
@@ -99,39 +83,20 @@ def __getattr__(name: str) -> object:
 
         return ServiceBase
 
-    # Services symbols (services/ may import from runtime)
-    if name == "ErrorTrackingService":
-        from vibe3.services.error_tracking_service import ErrorTrackingService
-
-        return ErrorTrackingService
-
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 
 __all__ = [
-    # Circuit breaker
     "CircuitBreaker",
     "CircuitState",
     "ErrorCategory",
     "classify_failure",
-    # Cleanup executor
     "execute_expired_resource_cleanup",
-    # Heartbeat server
     "HeartbeatServer",
-    # Periodic check executor
     "execute_periodic_check",
-    # Service protocol
     "ServiceBase",
-    # Orchestra instance management
     "OrchestraInstanceInfo",
     "read_instance_info",
     "validate_instance",
     "write_instance_info",
-    # External dependencies
-    "OrchestraConfig",
-    "PeriodicCheckConfig",
-    "append_orchestra_event",
-    "append_orchestra_run_separator",
-    # Services (via __getattr__)
-    "ErrorTrackingService",
 ]
