+++ src/vibe3/runtime/cleanup_executor.py
@@ -2,8 +2,8 @@
 
 from loguru import logger
 
-from vibe3.config.orchestra_config import PeriodicCheckConfig
-from vibe3.orchestra.logging import append_orchestra_event
+from vibe3.config import PeriodicCheckConfig
+from vibe3.orchestra import append_orchestra_event
 
 
 async def execute_expired_resource_cleanup(
@@ -20,9 +20,7 @@ async def execute_expired_resource_cleanup(
         tick_number: Current tick number (for logging)
     """
     # Delay imports to avoid circular dependencies
-    from vibe3.clients.git_client import GitClient
-    from vibe3.clients.github_client import GitHubClient
-    from vibe3.clients.sqlite_client import SQLiteClient
+    from vibe3.clients import GitClient, GitHubClient, SQLiteClient
     from vibe3.services.expired_resource_cleanup_service import (
         ExpiredResourceCleanupService,
     )
